feat: define task type to int64

+++ internal/tss/types.go
@@ -23,7 +23,7 @@ func convertMsgData(msg p2p.Message[json.RawMessage]) any {
 	case DataTypeTssKeygenMsg, DataTypeTssReSharingMsg, DataTypeTssSignMsg:
 		return unmarshal[types.TssMessage](msg.Data)
 	case GenKeySessionType, SignSessionType, ReShareGroupSessionType:
-		return unmarshal[SessionMessage[int32]](msg.Data)
+		return unmarshal[SessionMessage[int64]](msg.Data)
 	case DataTypeSignCreateWallet:
 		return unmarshal[types.SignMessage](msg.Data)
 	}
